Chức năng đăng ký ghế ngồi

+++ database/migrations/2018_11_29_132503_remove_trainer_id_column_in_users_table.php
@@ -13,11 +13,23 @@ class RemoveTrainerIdColumnInUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('trainer_id');
-            $table->dropColumn('seat_id');
-            $table->dropColumn('images');
-        });
+        if (Schema::hasColumn('users', 'trainer_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('trainer_id');
+            });
+        }
+
+        if (Schema::hasColumn('users', 'seat_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('seat_id');
+            });
+        }
+
+        if (Schema::hasColumn('users', 'images')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('images');
+            });
+        }
     }
 
     /**
